scripts/twop/requirements.txt
python-dotenv>=1.0
requests>=2.31
pyyaml>=6.0
